move data to utils
